+++ autonomous_agent_api/backend/app/models/trigger/resposne_dto.py
@@ -1,13 +1,15 @@
 from pydantic import BaseModel
 from typing import Union
 
-from backend.app.models.trigger.trigger_dto import CronTriggerDTO, TopicTriggerDTO
+
+from backend.app.models.trigger.trigger_dto import Action, CronTriggerDTO, TopicTriggerDTO
 
 
 class TriggerResponse(BaseModel):
     id: str
     agent_id: str
     type: str
+    action: Action
     data: Union[CronTriggerDTO, TopicTriggerDTO]