+++ autonomous_agent_api/backend/app/models/trigger/trigger_dto.py
@@ -1,6 +1,6 @@
 from fastapi import HTTPException
 from typing import Union
-from pydantic import BaseModel
+from pydantic import BaseModel, json
 from croniter import croniter
 
 
@@ -13,17 +13,26 @@ class TopicTriggerDTO(BaseModel):
     topic: str
 
 
+class Action(BaseModel):
+    function_name: str
+    parameter: list
+
+
 class TriggerCreateDTO(BaseModel):
     type: str
+    action: Action
     data: Union[CronTriggerDTO, TopicTriggerDTO]
 
 
 # validation for cron expression
-async def validate_type_CRON(cron_expression: str):
+async def validate_type_CRON(cron_expression: str, probability: float):
     try:
         croniter(cron_expression)
     except ValueError as e:
-        raise HTTPException(400, f"Invalid CRON expression: {str(e)}")
+        raise HTTPException(400, f"Invalid CRON expression")
+        # Validate probability
+    if probability < 0 or probability > 1:
+        raise HTTPException(400, "Probability must be between 0 and 1 (inclusive)")
 
 
 # validation for Topic