+++ autonomous_agent_api/prisma/migrations/20240412112012_initial/migration.sql
@@ -5,7 +5,8 @@ CREATE TYPE "TriggerType" AS ENUM ('CRON', 'TOPIC');
 CREATE TABLE "Agent" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
-    "action" TEXT[] DEFAULT ARRAY[]::TEXT[],
+    "instance" BIGINT NOT NULL,
+    "last_active" TIMESTAMP(3),
     "created_at" TIMESTAMP(3) NOT NULL,
     "updated_at" TIMESTAMP(3) NOT NULL,
     "deleted_at" TIMESTAMP(3),
@@ -18,6 +19,7 @@ CREATE TABLE "Trigger" (
     "id" TEXT NOT NULL,
     "agent_id" TEXT NOT NULL,
     "type" "TriggerType" NOT NULL,
+    "action" JSONB NOT NULL,
     "data" JSONB NOT NULL,
     "created_at" TIMESTAMP(3) NOT NULL,
     "updated_at" TIMESTAMP(3) NOT NULL,