+++ autonomous_agent_api/CHANGELOG.md
@@ -30,3 +30,22 @@
 
 # TITLE - AGENT Websocket , DATE- 2024-04-09
  - Added websocket endpoint for agent websocket connection
+
+# TITLE -  Agent registration and config fetch , DATE - 2024-04-09/10
+  - added instance and last_active field in Agent Field
+  - updated Agent Model and refactored repository 
+  - added features for fetching the configuration on the basis of agent_id
+  - Added functionality for periodically updating last_active for agent in agent table
+  - Added test case for the trigger repository
+
+ # TITLE -  Trigger Table update and define action in trigger config , DATE - 2024-04-11
+ - added action field in trigger table
+ - action field json 
+    - contains function_name and parameter as a details for action in dto model
+ - updated the trigger dto model and updated trigger respository class
+
+# TITLE -  Trigger Table update and define action in trigger config , DATE - 2024-04-12
+  - refactored the agent_websocket.py router class to websocket_router.py
+  - modularized the agent_websocket code into seprate service , repository , and controller class
+
+